mines & pmines performance fixes + tons of other fixes

+++ src/main/kotlin/net/evilblock/prisonaio/module/storage/command/WipeDatabaseCommand.kt
@@ -4,6 +4,7 @@ import net.evilblock.cubed.command.Command
 import net.evilblock.prisonaio.module.storage.StorageModule
 import org.bukkit.ChatColor
 import org.bukkit.command.CommandSender
+import org.bukkit.entity.Player
 
 object WipeDatabaseCommand {
 
@@ -14,6 +15,11 @@ object WipeDatabaseCommand {
     )
     @JvmStatic
     fun execute(sender: CommandSender) {
+        if (sender is Player) {
+            sender.sendMessage("${ChatColor.RED}That command must be executed from console!")
+            return
+        }
+
         StorageModule.database.drop()
         sender.sendMessage("${ChatColor.GREEN}Successfully wiped database!")
     }
